# server/shared-types/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tenant Leaf API
  version: 2.2.0-draft.1
  x-contract-status: review-draft
  description: |
    세입세잎 Android 앱과 Kotlin API 서버가 공유하는 MVP HTTP API 계약의 2.2 검토 초안입니다.
    현장 점검은 체크리스트 대신 구역·관찰·근거 JPEG를 중심으로 처리합니다.
    원본 영상은 휴대전화 갤러리에만 보관하고 서버에는 업로드하지 않습니다.
    AI 결과는 실제 하자나 안전 상태를 확정하지 않는 확인 필요 관찰 후보입니다.
    이 명세에는 인증, 매물, 임장, JPEG 업로드·집합 확정, 관찰 조회·검토와 자동 리포트 조회 API를 포함합니다.
servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
...
security:
  - bearerAuth: []
paths:
...
  /inspections/{inspectionId}/observations:
    parameters:
      - $ref: "#/components/parameters/InspectionId"
    get:
      tags: [Observations]
      summary: 임장의 확인 필요 관찰 목록 조회
      operationId: listInspectionObservations
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 관찰 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationPage"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /observations/{observationId}:
    parameters:
      - $ref: "#/components/parameters/ObservationId"
    get:
      tags: [Observations]
      summary: 확인 필요 관찰 상세 조회
      operationId: getObservation
      responses:
        "200":
          description: 관찰 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /observations/{observationId}/status:
    parameters:
      - $ref: "#/components/parameters/ObservationId"
    patch:
      tags: [Observations]
      summary: 관찰 검토 상태 변경
      description: AI 원본 결과는 유지하며 VIEWED 또는 DISMISSED로 변경합니다. DISMISSED 복원 값은 VIEWED입니다.
      operationId: updateObservationStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObservationStatusRequest"
      responses:
        "200":
          description: 관찰 상태 변경 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /properties/{propertyId}/reports:
    parameters:
      - $ref: "#/components/parameters/PropertyId"
    get:
      tags: [Reports]
      summary: 매물의 리포트 목록 조회
      operationId: listPropertyReports
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 리포트 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportPage"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /inspections/{inspectionId}/report:
    parameters:
      - $ref: "#/components/parameters/InspectionId"
    get:
      tags: [Reports]
      summary: 임장 자동 리포트 조회
      description: 미디어 집합 확정 전에는 리포트가 없을 수 있으며, 생성이 시작된 뒤에는 현재 상태와 사용 가능한 내용을 반환합니다.
      operationId: getInspectionReport
      responses:
        "200":
          description: 리포트 상태 또는 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-demo-token
  parameters:
    Page:
      name: page
      in: query
      description: 0부터 시작하는 페이지 번호
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: 페이지 크기
      schema:
        type: integer
        minimum: 1
        default: 20
    PropertyId:
      name: propertyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InspectionId:
      name: inspectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ObservationId:
      name: observationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 요청값 형식 또는 범위 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: 로그인이 필요함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 접근 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 요청한 리소스가 없거나 접근할 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 현재 상태에서는 요청을 수행할 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ObservationType:
      type: string
      description: AI classId 0~12와 순서대로 1:1 변환되는 비확정 확인 필요 관찰 유형입니다.
      enum:
        - CRACK_CHECK_NEEDED
        - MOLD_CHECK_NEEDED
        - PEELING_CHECK_NEEDED
        - WATER_DAMAGE_CHECK_NEEDED
        - TILE_DAMAGE_CHECK_NEEDED
        - HOLE_CHECK_NEEDED
        - TILE_CRACK_CHECK_NEEDED
        - PAINT_DRIPS_CHECK_NEEDED
        - PINHOLE_CHECK_NEEDED
        - SURFACE_DEFECT_CHECK_NEEDED
        - STAIN_CHECK_NEEDED
        - TROWEL_MARK_CHECK_NEEDED
        - OTHER_CHECK_NEEDED
    ImageDimensions:
      type: object
      required: [width, height]
      properties:
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
    BboxCoordinateSystem:
      type: string
      enum: [PIXEL_XYXY]
    EvidenceDetection:
      type: object
      required: [observationId, classId, label, displayLabel, displayColor, confidence, box]
      properties:
        observationId:
          type: string
          format: uuid
        classId:
          type: integer
          minimum: 0
          maximum: 12
        label:
          $ref: "#/components/schemas/AiLabel"
        displayLabel:
          type: string
          minLength: 1
        displayColor:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        box:
          $ref: "#/components/schemas/Bbox"
    ObservationEvidence:
      type: object
      required: [mediaId, primary, available, sourceVideoOffsetMs, image, coordinateSystem, detections]
      properties:
        mediaId:
          type: string
          format: uuid
        primary:
          type: boolean
          description: 관찰의 대표 근거이면 true입니다. 관찰마다 정확히 하나만 true입니다.
        available:
          type: boolean
          description: 보관 정책상 근거 JPEG를 현재 조회할 수 있는지 나타냅니다.
        sourceVideoOffsetMs:
          type: integer
          format: int64
          minimum: 0
        image:
          $ref: "#/components/schemas/ImageDimensions"
        coordinateSystem:
          $ref: "#/components/schemas/BboxCoordinateSystem"
        detections:
          type: array
          minItems: 1
          description: 같은 근거 JPEG에서 사용자 표시 임계값을 통과한 관찰 bbox입니다. 현재 관찰의 observationId와 일치하는 항목을 기본 강조합니다.
          items:
            $ref: "#/components/schemas/EvidenceDetection"
        viewUrl:
          type: string
          format: uri
          nullable: true
          description: 권한 확인 후 발급하는 짧은 만료 시간의 서명 조회 URL입니다.
        viewUrlExpiresAt:
          type: string
          format: date-time
          nullable: true
    Observation:
      type: object
      required:
        - id
        - inspectionId
        - type
        - zone
        - aiZone
        - zoneConfidence
        - zoneUncertain
        - zoneModelVersion
        - title
        - description
        - confidence
        - observationMinConfidence
        - status
        - aiClassId
        - aiLabel
        - modelVersion
        - evidence
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        inspectionId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ObservationType"
        zone:
          $ref: "#/components/schemas/Zone"
        aiZone:
          $ref: "#/components/schemas/Zone"
        zoneConfidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true
        zoneUncertain:
          type: boolean
        zoneModelVersion:
          type: string
          nullable: true
        userCorrectedZone:
          allOf:
            - $ref: "#/components/schemas/Zone"
          nullable: true
        correctedAt:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
          minLength: 1
          description: 하자를 확정하지 않는 규칙 기반 사용자 표시 제목입니다.
          example: 균열로 추정되는 흔적 확인 필요
        description:
          type: string
          minLength: 1
          description: 계약·안전·수리 필요를 단정하지 않는 규칙 기반 설명입니다.
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 원시 모델 신뢰도이며 위험도 또는 확률 보증이 아닙니다.
        observationMinConfidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 이 관찰 생성 당시 모델 버전·classId 설정에서 적용한 임계값입니다.
        status:
          $ref: "#/components/schemas/ObservationStatus"
        aiClassId:
          type: integer
          minimum: 0
          maximum: 12
        aiLabel:
          $ref: "#/components/schemas/AiLabel"
        modelVersion:
          type: string
          minLength: 1
        evidence:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: "#/components/schemas/ObservationEvidence"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ObservationPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Observation"
    UpdateObservationStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [VIEWED, DISMISSED]
          description: DISMISSED 관찰의 복원 요청은 VIEWED를 사용합니다.
    ReportFailureCode:
      type: string
      enum: [NO_ANALYZABLE_MEDIA, GENERATION_FAILED]
    ReportSummary:
      type: object
      required:
        - id
        - propertyId
        - inspectionId
        - status
        - successfulMediaCount
        - failedMediaCount
        - observationCount
        - propertyDisplayName
        - inspectionEndedAt
        - scoreIsProvisional
        - templateVersion
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
        inspectionId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ReportStatus"
        successfulMediaCount:
          type: integer
          minimum: 0
        failedMediaCount:
          type: integer
          minimum: 0
        observationCount:
          type: integer
          minimum: 0
          description: DISMISSED 관찰을 제외한 기본 리포트 관찰 수입니다.
        propertyDisplayName:
          type: string
          minLength: 1
        inspectionEndedAt:
          type: string
          format: date-time
        referenceScore:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: 확인 필요 관찰 1건당 5점을 차감한 참고값이며 안전·계약 점수가 아닙니다.
        scorePolicyVersion:
          type: string
          nullable: true
        scoreIsProvisional:
          type: boolean
          description: 일부 사진 분석 실패로 점수가 잠정값이면 true입니다.
        failureCode:
          allOf:
            - $ref: "#/components/schemas/ReportFailureCode"
          nullable: true
        templateVersion:
          type: string
          minLength: 1
          description: 자유 생성 LLM이 아닌 규칙·문장 템플릿 버전입니다.
        generatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/ReportSummary"
        - type: object
          required: [observations]
          properties:
            emptyObservationMessage:
              type: string
              nullable: true
              description: 관찰 0건 성공 시 '현재 촬영 근거에서 확인 필요 관찰이 생성되지 않았습니다.'를 제공합니다.
            observations:
              type: array
              description: DISMISSED 관찰은 제외하며 각 항목의 근거 mediaId로 문장을 추적할 수 있습니다.
              items:
                $ref: "#/components/schemas/Observation"
    ReportPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ReportSummary"
    Zone:
      type: string
      enum: [ENTRANCE_COMMON, KITCHEN, WINDOW_VENTILATION, LIVING_ROOM, BATHROOM, UNKNOWN]
...
    ObservationStatus:
      type: string
      description: DISMISSED는 하자 아님이나 오탐 확정이 아니라 기본 리포트 제외 상태입니다.
      enum: [ACTIVE, VIEWED, DISMISSED]
    ReportStatus:
      type: string
      enum: [NOT_REQUESTED, WAITING_FOR_ANALYSIS, GENERATING, COMPLETED, PARTIAL_COMPLETED, FAILED]
    AiLabel:
      type: string
      description: 모델 클래스 ID 0~12와 고정 쌍으로 사용하는 원본 영문 라벨입니다.
      enum:
        - crack
        - mold
        - peeling
        - waterdamage
        - tiledamage
        - hole
        - tilecrack
        - paintdrips
        - pinhole
        - surfacedefect
        - stain
        - trowelmark
        - other
    Bbox:
      type: object
      required: [left, top, right, bottom]
      description: EXIF 방향 정규화가 끝난 원본 전체 JPEG의 왼쪽 위를 원점으로 하는 픽셀 xyxy 좌표입니다.
      properties:
        left:
          type: number
          format: double
          minimum: 0
        top:
          type: number
          format: double
          minimum: 0
        right:
          type: number
          format: double
          minimum: 0
        bottom:
          type: number
          format: double
          minimum: 0
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    FieldError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: 요청값을 확인해 주세요.
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        traceId:
          type: string
          minLength: 1
          example: 7ccf5b3f44d94fa0